requirements.txt
selenium>=4.0.0
webdriver-manager>=3.8.0
yt-dlp>=2023.0.0
python-dotenv>=1.0.0
undetected-chromedriver>=3.5.0